requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Django==2.2.5
django-enumfield==1.5
django-environ==0.4.5
django-extensions==2.2.1
django-rest-framework==0.1.0
djangorestframework==3.10.3
factory-boy==2.12.0
Faker==2.0.2
importlib-metadata==0.23
invoke==1.3.0
more-itertools==7.2.0
packaging==19.1
pkg-resources==0.0.0
pluggy==0.13.0
psycopg2==2.8.3
py==1.8.0
pyparsing==2.4.2
pytest==5.1.2
pytest-django==3.5.1
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
PyYAML==5.1.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
uritemplate==3.0.0
wcwidth==0.1.7
zipp==0.6.0
